renamed codingame package
