+++ src/main/scala/com/truelaurel/codingame/challenge/GameLoop.scala
@@ -1,4 +1,4 @@
-package com.truelaurel.codingame.architecture
+package com.truelaurel.codingame.challenge
 
 import com.truelaurel.codingame.tool.logging.CGLogger
