+++ src/main/scala/com/truelaurel/codingame/challenge/GameResult.scala
@@ -1,4 +1,4 @@
-package com.truelaurel.codingame.architecture
+package com.truelaurel.codingame.challenge
 
 sealed trait GameResult
